+++ src/Symfony/Bridge/Twig/Command/LintCommand.php
@@ -17,6 +17,8 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Filesystem\Watcher\FileChangeEvent;
 use Symfony\Component\Finder\Finder;
 use Twig\Environment;
 use Twig\Error\Error;
@@ -47,6 +49,7 @@ class LintCommand extends Command
         $this
             ->setDescription('Lints a template and outputs encountered errors')
             ->addOption('format', null, InputOption::VALUE_REQUIRED, 'The output format', 'txt')
+            ->addOption('watch', null, InputOption::VALUE_NONE, 'Watch the files or directories for changes')
             ->addArgument('filename', InputArgument::IS_ARRAY)
             ->setHelp(<<<'EOF'
 The <info>%command.name%</info> command lints a template and outputs to STDOUT
@@ -90,7 +93,24 @@ EOF
 
         $filesInfo = $this->getFilesInfo($filenames);
 
-        return $this->display($input, $output, $io, $filesInfo);
+        $display = $this->display($input, $output, $io, $filesInfo);
+
+        if (!$input->getOption('watch')) {
+            return $display;
+        }
+
+        if (!class_exists(Filesystem::class)) {
+            throw new \RuntimeException('The symfony/filesystem package is required in order to watch files for changes.');
+        }
+
+        $io->note('Watching files for changes...');
+
+        $fs = new Filesystem();
+        $fs->watch($filenames, function (string $file, int $event) use ($io, $input, $output) {
+            if (FileChangeEvent::FILE_DELETED !== $event) {
+                $this->display($input, $output, $io, array($this->validate(file_get_contents($file), $file)));
+            }
+        });
     }
 
     private function getFilesInfo(array $filenames)